refactor: 💡 split queries.ts
