+++ src/middleware/queries/index.ts
@@ -1,0 +1,4 @@
+/* istanbul ignore file */
+
+export * from './regular';
+export * from './confidential';
